requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
cryptography==3.4.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django>=3.1.8
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-recaptcha==2.0.6
django-username-email==2.4.1
django-widget-tweaks==1.4.8
Faker==7.0.1
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0